Remove all files
NestJS was used as it seemed like a good framework for supporting REST
API development with suitable concepts like interceptors, pipes and
guards. However, NestJS lacks sufficient documentation and community,
making it very difficult to try to find information on NestJS.

The plan is to abort any progress with NestJS and move on to express.js
with the latter having much better documentation and community support.